+++ R/accessors-second.r
@@ -59,10 +59,10 @@ setMethod("second<-", "Interval", function(x, value) {
 
 #' @export
 setMethod("second<-", "POSIXt", function(x, value) {
-  update.POSIXt(x, seconds = value)
+  update_datetime(x, seconds = value, roll_dst = "NA")
 })
 
 #' @export
 setMethod("second<-", "Date", function(x, value) {
-  update.Date(x, seconds = value)
+  update_datetime(x, seconds = value, roll_dst = "NA")
 })