+++ R/ops-addition.r
@@ -53,21 +53,23 @@ add_period_to_date <- function(per, date) {
   lt <- add_months(lt, ms)
 
   if (is.Date(date)) {
-    new <- update(as.Date(lt),
-      days = mday(lt) + per@day,
-      hours = per@hour,
-      minutes = per@minute,
-      seconds = per@.Data
-    )
+    new <-
+      update.Date(as.Date(lt),
+        days = mday(lt) + per@day,
+        hours = per@hour,
+        minutes = per@minute,
+        seconds = per@.Data
+      )
     return(new)
   }
 
-  new <- update(lt,
-    days = mday(lt) + per@day,
-    hours = hour(lt) + per@hour,
-    minutes = minute(lt) + per@minute,
-    seconds = second(lt) + per@.Data
-  )
+  new <-
+    update.POSIXt(lt,
+      days = mday(lt) + per@day,
+      hours = hour(lt) + per@hour,
+      minutes = minute(lt) + per@minute,
+      seconds = second(lt) + per@.Data
+    )
 
   reclass_date(new, date)
 }